release: build for macOS and Windows separately
Update the GitHub Actions 'Release' workflow to build for macOS and
Windows within separate jobs, so we perform the macOS builds on a macOS
runner and the Windows builds on a Windows runner. This is in
preparation for building against OS-specific libraries (with cgo).

Continue to use GoReleaser for the Linux and FreeBSD builds, as well as
the Linux packages and the Docker images.

Add a separate job at the beginning of the workflow for setting version
flags to be shared across all three build jobs, and a separate job at
the end for uploading all of the release asset checksums in one file.