integrate with system cert store (#314)
Add a new package 'certstore' to provide a common interface for the
macOS and Windows platform-specific trust store methods from the
vendored 'third_party/ecpsigner' code.

Add a new --client-cert-issuer-cn flag to the tcp command using this
functionality. Register this flag only when the feature is supported.

Update the version output to add a new 'Features:' line to indicate
whether pomerium-cli was compiled with support for this feature. Give
the macOS integration the feature name 'keychain' and the Windows
integration the feature name 'ncrypt'.