expand client cert search capabilities
Replace the --client-cert-issuer-cn option with a new set of options:

  --client-cert-from-store  enables searching the OS cert store

  --client-cert-issuer      selects between multiple available
certificates based on an attribute of the cert's Issuer name

  --client-cert-subject     selects between multiple available
certificates based on an attribute of the cert's Subject name

If only the first option is set, the cert store will be searched based
on any CA names provided in the TLS Certificate Request message.

Set a finalizer on the Windows credential type to release OS resources
automatically. (The macOS credential type already does this.)