ci: rework homebrew-tap updates (#519)
The macOS and Windows binaries are currently built separately from 
GoReleaser (in order to be able to build against OS-native libraries
using cgo). As a result, GoReleaser builds only the Linux binaries, and
so it will update the Homebrew tap to be Linux-only.

Remove the 'brews' section from the GoReleaser configuration, and 
instead trigger the replacement repository_dispatch workflow in the
pomerium/homebrew-tap repo.